moved nested components out of delivery into individual files
