+++ public/components/create_report/report_settings/report_settings.js
@@ -45,7 +45,7 @@ const ReportSettings = (props) => {
           <EuiFlexItem>
             <EuiFormRow label="Name" helpText="Specify a descriptive report name">
               <EuiFieldText
-                placeholder="Report Name"
+                placeholder="Report name"
               />
             </EuiFormRow>
           </EuiFlexItem>
